requirements.txt
Django==4.2.6
django-environ==0.11.2
requests==2.31.0
python-dateutil==2.8.2

psycopg2==2.9.9